requirements.txt
Django
mysqlclient
djangorestframework
django-rest-swagger
drf-yasg
django-rest-registration
django-cors-headers